llm-server/requirements.txt
fastapi==0.108.0
uvicorn[standard]==0.25.0
llama-cpp-python==0.2.27
pydantic==2.5.3
huggingface-hub==0.20.1
